feat: polsih error refactor Because:
* We had some last minute tweaks to the error arrangement.

This commit:

* Moves the error class to the library using it and inherits from a
  shared base error.
