+++ packages/fxa-auth-server/lib/payments/paypal/processor.ts
@@ -6,6 +6,7 @@ import { Logger } from 'mozlog';
 import Stripe from 'stripe';
 import { Container } from 'typedi';
 
+import { PayPalClientError } from '../../../../../libs/payments/paypal/src';
 import { ConfigType } from '../../../config';
 import error from '../../error';
 import { StripeWebhookHandler } from '../../routes/subscriptions/stripe-webhook';
@@ -17,7 +18,6 @@ import {
   PAYPAL_SOURCE_ERRORS,
 } from './error-codes';
 import { PayPalHelper, TransactionSearchResult } from './helper';
-import { PayPalClientError } from '../../../../../libs/shared/error/src';
 
 /**
  * Generest a timestamp in seconds that is `hours` before the current
